requirements.txt
PyYAML
streamlit
streamlit-authenticator
streamlit-server-state
streamlit-sortables
